+++ Changelog.md
@@ -1,6 +1,24 @@
 ### 🌌 Next Version
 The future is uncertain
 ___
+### 4.1.5
+Added Power bars to the Border Auto Adjust function  
+Added borders to Experience and Reputation databars  
+Added Totem Item Level font to the Font functions  
+Added borders to Alternative Power bar  
+Added zhTW translation by 魔能機曱  
+Added a few missing shadows  
+Updated Borders by Class Color to use the Power Colors for Player and Target, Target of Target doesn't have events that fire to update it  
+Updated Dark Souls death animation to better position the text  
+Updated Inspect Item Level text in Retail to fix the font size  
+Updated some profiles to use general ElvUI font outline  
+Updated LFG skin to desaturate when unavailable  
+Updated German locale by Dlarge  
+Fixed an issue with nameplate Threat colors where Off Tanks didnt get the correct color (thanks Trenchy)  
+Fixed an issue where non Fire Mages could get the Nameplate Powerbar stuck in wrong units  
+Fixed an issue due to MerathilisUI removing some of its options  
+Fixed Inspect Talent Size in Cataclysm  
+Fixed Retail World Map shadows  
 ### 4.1.4
 Added Hunter Pet Stable in Cataclysm since it now uses the old Retail Stables  
 Added backgrounds to Portrait Skin